+++ bika/lims/catalog/__init__.py
@@ -1,0 +1,16 @@
+# This file is part of Bika LIMS
+#
+# Copyright 2011-2016 by it's authors.
+# Some rights reserved. See LICENSE.txt, AUTHORS.txt.
+
+# Catalog IDs static constant
+from .analysisrequest_catalog import CATALOG_ANALYSIS_REQUEST_LISTING
+# Catalog classes
+from .bika_catalog import BikaCatalog
+from .bikasetup_catalog import BikaSetupCatalog
+from .analysis_catalog import BikaAnalysisCatalog
+from .analysisrequest_catalog import BikaCatalogAnalysisRequestListing
+# Catalog public functions
+from .catalog_utilities import getCatalogDefinitions
+from .catalog_utilities import setup_catalogs
+from .catalog_utilities import getCatalog
