+++ bika/lims/catalog/bika_catalog.py
@@ -1,0 +1,65 @@
+# This file is part of Bika LIMS
+#
+# Copyright 2011-2016 by it's authors.
+# Some rights reserved. See LICENSE.txt, AUTHORS.txt.
+
+import sys
+import traceback
+from zope.interface import implements
+from AccessControl import ClassSecurityInfo
+from App.class_init import InitializeClass
+from Products.CMFCore.permissions import ManagePortal
+from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.CatalogTool import CatalogTool
+from Products.ZCatalog.ZCatalog import ZCatalog
+# Bika LIMS imports
+from bika.lims import logger
+from bika.lims.interfaces import IBikaCatalog
+
+
+class BikaCatalog(CatalogTool):
+
+    """Catalog for various transactional types"""
+
+    implements(IBikaCatalog)
+
+    security = ClassSecurityInfo()
+    _properties = ({'id': 'title', 'type': 'string', 'mode': 'w'},)
+
+    title = 'Bika Catalog'
+    id = 'bika_catalog'
+    portal_type = meta_type = 'BikaCatalog'
+    plone_tool = 1
+
+    def __init__(self):
+        ZCatalog.__init__(self, self.id)
+
+    security.declareProtected(ManagePortal, 'clearFindAndRebuild')
+
+    def clearFindAndRebuild(self):
+        """
+        """
+
+        def indexObject(obj, path):
+            self.reindexObject(obj)
+        logger.info('Cleaning and rebuilding %s...' % self.id)
+        try:
+            at = getToolByName(self, 'archetype_tool')
+            types = [k for k, v in at.catalog_map.items()
+                     if self.id in v]
+
+            self.manage_catalogClear()
+            portal = getToolByName(self, 'portal_url').getPortalObject()
+            portal.ZopeFindAndApply(portal,
+                                    obj_metatypes=types,
+                                    search_sub=True,
+                                    apply_func=indexObject)
+        except:
+            logger.error(traceback.format_exc())
+            e = sys.exc_info()
+            logger.error(
+                "Unable to clean and rebuild %s due to: %s" % (self.id, e))
+        logger.info('%s cleaned and rebuilt' % self.id)
+
+
+InitializeClass(BikaCatalog)
